Create U1HW1.java
For some reason my '\r' is not letting me edit lines like I thought it would, so you may notice that my addAll method goes to a new line after each part of the equation while printing because it is a different print statement for each number. I was thinking about making an array, adding all the different numbers with the plus signs into the array, and then concentrating the different elements into one string so I could make the equation in one line.